requirements.txt
psycopg2-binary>=2.9
requests>=2.31
beautifulsoup4>=4.12
rapidfuzz>=3.0
python-dotenv>=1.0
